renamed prod_neuron to prodneuron. Docstrings.
